requirements.txt
django==5.0.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
django-cors-headers==4.3.1
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
zk==0.9.4
reportlab==4.0.9
psycopg[binary]==3.3.2
openpyxl==3.1.5
